refactor: ダッシュボードのユーティリティ関数をutils.tsに分離
テスト対象のロジック（formatDate, getPresetLabel, STATUS_CONFIG,
PLAN_LABELS, calculateUsagePercentage）を utils.ts に集約し、
表示専用の .tsx ファイルはカバレッジ除外に変更。
JSXの条件分岐はE2Eテストで検証する方針に統一。
