+++ packages/web/src/components/dashboard/report-history.tsx
@@ -1,5 +1,4 @@
 import Link from "next/link";
-import { findPreset } from "@townlens/core";
 import {
   Table,
   TableBody,
@@ -9,6 +8,7 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import { Badge } from "@/components/ui/badge";
+import { formatDate, getPresetLabel, STATUS_CONFIG } from "./utils";
 
 interface ReportHistoryProps {
   readonly reports: ReadonlyArray<{
@@ -19,28 +19,6 @@ interface ReportHistoryProps {
     readonly created_at: string;
   }>;
 }
-
-/** ISO 8601 日時文字列を "YYYY/MM/DD HH:mm" にフォーマット */
-export function formatDate(isoString: string): string {
-  return new Intl.DateTimeFormat("ja-JP", {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-    hour: "2-digit",
-    minute: "2-digit",
-  }).format(new Date(isoString));
-}
-
-/** プリセット名をラベルに変換（見つからない場合は生の名前を返す） */
-export function getPresetLabel(presetName: string): string {
-  return findPreset(presetName)?.label ?? presetName;
-}
-
-export const STATUS_CONFIG = {
-  completed: { label: "完了", variant: "default" as const },
-  processing: { label: "生成中", variant: "secondary" as const },
-  failed: { label: "失敗", variant: "destructive" as const },
-};
 
 /** レポート履歴テーブル */
 export function ReportHistory({ reports }: ReportHistoryProps) {
